feat[wip]: adds new method of reverse geocoding
This largely accomplishes two things:
  * Addresses Issue #33 (country borders not being respected)
  * Makes changing data sources more modular

The dataset being used for country boundaries comes from
[GADM](https://gadm.org/). Unfortunately, while they do have a global
ADM_0 set, it's only available as a layer in a single GeoPackage which
is 2.6 GB, and thus also requires initial extraction / conversion with
ogr2ogr to be useful here.

The geonames dataset is replaced with the finer-resolution version, as
the original one was missing several cities in testing.

The new query for finding a match uses the `&&` PostGIS operator for the
geometries, which returns True for any intersection. This method is an
order of magnitude faster (~2 msec vs. ~42 msec in my tests) than
Voronoi and ST_CONTAINS, and does not seem to suffer from errors that I
can tell (see included additional tests), but it does heavily rely on
the source dataset being accurate. Interestingly, using the `geom`
columns on their own as a comparison with ST_CONTAINS _does_ result in
erroneous results.